fix: resolve decoding edge cases, add code signing and update workflow
- Fix file name truncation bug in strip_prefix logic
- Fix Blowfish fallback payload offset in Ver413
- Fix false positive BM header matches in Ver121
- Optimize RSA BigUint allocations in decoding loop
- Fix 32-bit usize file size truncation in scan module
- Add RSA profile odd modulus validation
- Integrate digital signing for binary and proxy DLL into release CI

+++ src/storage/scan.rs
@@ -45,7 +45,7 @@ pub fn scan_tree(root: &Path, on_progress: &mut dyn FnMut(usize)) -> ScanResult 
             let path = entry.path();
             let size = entry.metadata().map(|m| m.len()).unwrap_or(0);
 
-            if size as usize > aac::PAYLOAD_OFFSET
+            if size > aac::PAYLOAD_OFFSET as u64
                 && let Some(header) = read_header(path, 20)
                 && aac::header_is_aac(&header)
             {
